pyproject.toml
[project]
name = "roku-app-build"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.1.1",
    "pytablericons>=1.0.1",
    "requests",
]

[project.scripts]
build = "build:main"
deploy = "deploy:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["build.py", "deploy.py"]
